+++ src/main/kotlin/icu/windea/pls/script/psi/ParadoxScriptPsiSearchExtensions.kt
@@ -211,9 +211,9 @@ fun <T : ParadoxScriptMemberElement> ParadoxScriptMemberElement.findByPath(
     var current: ParadoxScriptMemberElement = this
     if(path.isNotEmpty()) {
         val elementPath = ParadoxElementPath.resolve(path)
-        val subPathInfos = elementPath.subPathInfos
+        val subPathInfos = elementPath.subPaths
         for(subPathInfo in subPathInfos) {
-            val (subPath) = subPathInfo
+            val subPath = subPathInfo.subPath
             if(subPath == "-") return null //TODO 暂不支持
             current = current.findProperty(subPath, ignoreCase, conditional, inline) ?: return null
         }
@@ -308,9 +308,9 @@ fun ParadoxScriptMemberElement.findParentByPath(
     var current: ParadoxScriptMemberElement = this
     if(path.isNotEmpty()) {
         val elementPath = ParadoxElementPath.resolve(path)
-        val subPathInfos = elementPath.subPathInfos
+        val subPathInfos = elementPath.subPaths
         for(subPathInfo in subPathInfos.reversed()) {
-            val (subPath) = subPathInfo
+            val subPath = subPathInfo.subPath
             if(subPath == "-") return null //TODO 暂不支持
             current = current.findParentProperty(subPath, ignoreCase, link = link) ?: return null
         }