+++ src/main/kotlin/icu/windea/pls/core/index/ParadoxValueSetValueIndex.kt
@@ -116,3 +116,4 @@ object ParadoxValueSetValueIndex {
         return gist.getFileData(project, file)
     }
 }
+